Fix build on macOS (#282)
* Avoid redeclaring std::hash and std::less.

These duplicate declarations (in addition to the ones in the standard
library) caused problems when compiling on macOS.

* Do not build testPcieErrorHandling when PCIe backend is not available.

* Use correct data type when calling NumericAddressedBackend::read.

There are two versions of NumericAddressedBackend::read, and neither of
them expects values of type size_t for the first two arguments, so
calling it with such values results in an ambiguity which version should
be used.

By making sure that we call it with arguments of type uint64_t, we can
resolve this ambiguity.

+++ tests/executables_src/testNumericAddressedBackendRegisterAccessor.cpp
@@ -25,7 +25,7 @@ using namespace ChimeraTK;
 // we need to access some private data of the low level transfer element
 // FIXME: redesign test so this is not necessary!
 struct NumericAddressedLowLevelTransferElement_startAddress {
-  typedef size_t NumericAddressedLowLevelTransferElement::*type;
+  typedef uint64_t NumericAddressedLowLevelTransferElement::*type;
 };
 template struct accessPrivateData::stow_private<NumericAddressedLowLevelTransferElement_startAddress,
     &ChimeraTK::NumericAddressedLowLevelTransferElement::_startAddress>;
